+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { theme } from './theme';
 import { RecoilRoot } from 'recoil';
 import './index.css';
-import SignUp from './SignUp.tsx';
+import SignUp from './pages/SignUp.tsx';
 
 const GlobalStyle = createGlobalStyle`
 @import url('https://fonts.googleapis.com/css2?family=Source+Serif+4:ital,opsz,wght@1,8..60,300&display=swap');
