+++ frontend/src/index.css
@@ -6,20 +6,20 @@
 @font-face {
     font-family: 'Pre-R';
     src: local('Pre-R'),
-    url('./fonts/Pretendard/Pretendard-Regular.woff');
+    url('shared/fonts/Pretendard/Pretendard-Regular.woff');
     font-weight: 400;
 }
 
 @font-face {
     font-family: 'Pre-R';
     src: local('Pre-R'),
-    url('./fonts/Pretendard/Pretendard-Regular.woff');
+    url('shared/fonts/Pretendard/Pretendard-Regular.woff');
     font-weight: 400;
 }
 
 @font-face {
     font-family: 'Bayon';
     src: local('Bayon'),
-    url('./fonts/Bayon/Bayon-Regular.ttf');
+    url('shared/fonts/Bayon/Bayon-Regular.ttf');
     font-weight: 400;
 }
