+++ frontend/src/pages/SignUp.tsx
@@ -2,7 +2,7 @@ export default function SignUp() {
   return (
     //   확인용 배경 박스
     <div className="bg-[#111111] bg-opacity-90 flex w-screen h-screen justify-center items-center">
-      <div className="rounded-[20px] bg-[url('./SignUpImg/signup.svg')] bg-cover bg-no-repeat flex flex-row w-[960px] h-[600px] box-sizing-border">
+      <div className="rounded-[20px] bg-[url('./shared/assests/signup.svg')] bg-cover bg-no-repeat flex flex-row w-[960px] h-[600px] box-sizing-border">
         {/*왼쪽 박스*/}
         <div className="flex flex-col justify-center ml-8 box-sizing-border w-[55%] gap-2">
           <p className="break-words font-['Bayon'] font-normal text-[3rem] text-[#FFFFFF]">Techeer.log</p>
