--- frontend/src/styled.d.ts
@@ -1,9 +1,0 @@
-import 'styled-components';
-
-declare module 'styled-components' {
-  export interface DefaultTheme {
-    textColor: string;
-    bgColor: string;
-    accentColor: string;
-  }
-}
